requirement.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
langchain_google_genai
google-generativeai